MUI Autocomplete component for tagIds/sections fields - initial work completed
