+++ public/src/components/channelManagement/epicTests/epicTestTargetContentEditor.tsx
@@ -1,9 +1,7 @@
-import React, { useEffect, useState } from 'react';
-import { useForm } from 'react-hook-form';
+import React, { useEffect, useState, ChangeEvent } from 'react';
 import { Theme, makeStyles, TextField } from '@material-ui/core';
 
 import { Autocomplete } from '@material-ui/lab';
-import { SectionDataOrEmpty, TagIdDataOrEmpty, tagIdOptions, sectionOptions, callCAPI } from './fetchDataFromCapi';
 
 const useStyles = makeStyles(({ spacing }: Theme) => ({
   container: {
@@ -13,8 +11,6 @@ const useStyles = makeStyles(({ spacing }: Theme) => ({
     },
   },
 }));
-
-type StringDataOrEmpty = string[] | [];
 
 interface FormData {
   tagIds: string[];
@@ -37,6 +33,50 @@ interface EpicTestTargetContentEditorProps {
   ) => void;
 }
 
+interface TagIdData {
+  id: string;
+  display: string;
+  section: string;
+}
+
+interface SectionData {
+  id: string;
+  display: string;
+}
+
+interface RawTagIdData {
+  id?: string;
+  type?: string;
+  sectionId?: string;
+  sectionName?: string;
+  webTitle?: string;
+  webUrl?: string;
+  apiUrl?: string;
+  description?: string;
+  internalName?: string;
+}
+
+interface RawSectionsEditionData {
+  id?: string;
+  webTitle: string;
+  webUrl?: string;
+  apiUrl?: string;
+  code?: string;
+}
+
+interface RawSectionData {
+  id?: string;
+  webTitle: string;
+  webUrl?: string;
+  apiUrl?: string;
+  code?: string;
+  editions?: RawSectionsEditionData[];
+}
+
+type SectionDataOrEmpty = SectionData[] | [];
+type TagIdDataOrEmpty = TagIdData[] | [];
+type StringDataOrEmpty = string[] | [];
+
 const EpicTestTargetContentEditor: React.FC<EpicTestTargetContentEditorProps> = ({
   tagIds,
   sections,
@@ -45,21 +85,61 @@ const EpicTestTargetContentEditor: React.FC<EpicTestTargetContentEditorProps> = 
   editMode,
   updateTargetContent,
 }: EpicTestTargetContentEditorProps) => {
-
-  callCAPI();
-
   const classes = useStyles();
 
-  const onSubmit = (): void => {
+  // Infrastructure for tagId selectors
+  const [tagIdsOptions, setTagIdsOptions] = useState<TagIdDataOrEmpty>([]);
+  const [tagIdsField, setTagIdsField] = useState<TagIdDataOrEmpty>([]);
+  const [excludeTagIdsField, setExcludeTagIdsField] = useState<TagIdDataOrEmpty>([]);
 
-    updateTargetContent(
-      prepareTagIdsForSubmit(tagIdsField),
-      prepareSectionsForSubmit(sectionsField),
-      prepareTagIdsForSubmit(excludeTagIdsField),
-      prepareSectionsForSubmit(excludeSectionsField),
-    );
+  const getInitialTagIds = (vals: string[]): TagIdDataOrEmpty => {
+    const res = [];
+
+    for (const t of tagIdsOptions) {
+      if (vals.indexOf(t.id) >= 0) {
+        res.push(t);
+      }
+    }
+    return res;
   };
 
+  useEffect(() => {
+    console.log('tagIdsOptions length', tagIdsOptions.length);
+    setTagIdsField(getInitialTagIds(tagIds));
+    setExcludeTagIdsField(getInitialTagIds(excludeTagIds));
+  }, [tagIdsOptions]);
+
+  useEffect(() => setTagIdsField(getInitialTagIds(tagIds)), [tagIds]);
+
+  useEffect(() => setExcludeTagIdsField(getInitialTagIds(excludeTagIds)), [excludeTagIds]);
+
+  // Infrastructure for sections selectors
+  const [sectionsOptions, setSectionsOptions] = useState<SectionDataOrEmpty>([]);
+  const [sectionsField, setSectionsField] = useState<SectionDataOrEmpty>([]);
+  const [excludeSectionsField, setExcludeSectionsField] = useState<SectionDataOrEmpty>([]);
+
+  const getInitialSections = (vals: string[]): SectionDataOrEmpty => {
+    const res = [];
+
+    for (const s of sectionsOptions) {
+      if (vals.indexOf(s.id) >= 0) {
+        res.push(s);
+      }
+    }
+    return res;
+  };
+
+  useEffect(() => {
+    console.log('sectionsOptions length', sectionsOptions.length);
+    setSectionsField(getInitialSections(sections));
+    setExcludeSectionsField(getInitialSections(excludeSections));
+  }, [sectionsOptions]);
+
+  useEffect(() => setSectionsField(getInitialSections(sections)), [sections]);
+
+  useEffect(() => setExcludeSectionsField(getInitialSections(excludeSections)), [excludeSections]);
+
+  // Processing required to update tagId/section data immediately before submission
   const prepareTagIdsForSubmit = (input: TagIdDataOrEmpty): StringDataOrEmpty => {
     const output: string[] = [];
     input.forEach(i => output.push(i.id));
@@ -72,68 +152,180 @@ const EpicTestTargetContentEditor: React.FC<EpicTestTargetContentEditorProps> = 
     return output;
   };
 
-  // tagId stuff
-  const getInitialTagIds = (vals: string[]): TagIdDataOrEmpty => {
-
-    const res = [];
-
-    for (let t of tagIdOptions) {
-
-      if (vals.indexOf(t.id) >= 0) {
-        res.push(t);
-      }
-    };
-    return res;
+  const onSubmit = (): void => {
+    updateTargetContent(
+      prepareTagIdsForSubmit(tagIdsField),
+      prepareSectionsForSubmit(sectionsField),
+      prepareTagIdsForSubmit(excludeTagIdsField),
+      prepareSectionsForSubmit(excludeSectionsField),
+    );
   };
 
-  const [tagIdsField, setTagIdsField] = useState<TagIdDataOrEmpty>([]);
+  // Fetch data from CAPI
+  const [capiFlag, setCapiFlag] = useState(false);
 
-  const [excludeTagIdsField, setExcludeTagIdsField] = useState<TagIdDataOrEmpty>([]);
-  
-  useEffect(() => {
-    setTagIdsField(getInitialTagIds(tagIds));
-    setExcludeTagIdsField(getInitialTagIds(excludeTagIds));
-  }, [tagIdOptions]);
+  const rawTagIdResults: RawTagIdData[] = []; // Not stateful
 
-  useEffect(() => setTagIdsField(getInitialTagIds(tagIds)), [tagIds]);
+  const processTagIdResults = (data: RawTagIdData[] = []): void => {
+    const res: TagIdData[] = [];
 
-  useEffect(() => setExcludeTagIdsField(getInitialTagIds(excludeTagIds)), [excludeTagIds]);
+    if (data.length) {
+      data.forEach(item => {
+        const { id, webTitle, sectionName } = item;
 
-  // section stuff
-  const getInitialSections = (vals: string[]): SectionDataOrEmpty => {
+        if (id && webTitle && sectionName) {
+          const obj: TagIdData = {
+            id: id,
+            display: `${webTitle} [${id}]`,
+            section: sectionName.toUpperCase(),
+          };
+          res.push(obj);
+        }
+      });
 
-    const res = [];
-
-    for (let s of sectionOptions) {
-
-      if (vals.indexOf(s.id) >= 0) {
-        res.push(s);
-      }
-    };
-    return res;
+      res.sort((a, b) => {
+        if (a.section > b.section) {
+          return 1;
+        } else if (a.section < b.section) {
+          return -1;
+        } else {
+          if (a.display > b.display) {
+            return 1;
+          }
+          return -1;
+        }
+      });
+    }
+    setTagIdsOptions(res);
   };
 
-  const [sectionsField, setSectionsField] = useState<SectionDataOrEmpty>([]);
-  
-  const [excludeSectionsField, setExcludeSectionsField] = useState<SectionDataOrEmpty>([]);
+  const processSectionResults = (data: RawSectionData[] = []): void => {
+    const res: SectionData[] = [];
 
-  useEffect(() => {
-    setSectionsField(getInitialSections(sections));
-    setExcludeSectionsField(getInitialSections(excludeSections));
-  }, [sectionOptions])
+    if (data.length) {
+      data.forEach(item => {
+        const { id, webTitle } = item;
 
-  useEffect(() => setSectionsField(getInitialSections(sections)), [sections]);
+        if (id && webTitle) {
+          const obj: SectionData = {
+            id: id,
+            display: webTitle,
+          };
+          res.push(obj);
+        }
+      });
 
-  useEffect(() => setExcludeSectionsField(getInitialSections(excludeSections)), [excludeSections]);
+      res.sort((a, b) => {
+        if (a.display > b.display) {
+          return 1;
+        }
+        return -1;
+      });
+    }
+    setSectionsOptions(res);
+  };
 
+  const generateFetchPromise = (
+    url: string,
+    promises: Promise<RawTagIdData | RawSectionData>[],
+  ): Promise<RawTagIdData | RawSectionData> => {
+    return new Promise((resolve, reject) => {
+      fetch(url)
+        .then(res => res.json())
+        .then(res => {
+          if (res && res.response) {
+            const response = res.response;
+
+            if (response.currentPage < response.pages) {
+              url = url.replace(`&page=${response.currentPage}`, '');
+              url += `&page=${response.currentPage + 1}`;
+              promises.push(generateFetchPromise(url, promises));
+            }
+
+            if (response.results && response.results.length) {
+              rawTagIdResults.push(...response.results);
+            }
+          }
+          resolve(res);
+        })
+        .catch(e => reject(e));
+    });
+  };
+
+  const callCAPI = (): void => {
+    if (!capiFlag) {
+      console.log('Starting CAPI fetch');
+
+      setCapiFlag(true);
+
+      const sectionUrl = `${window.location.origin}/capi/sections?page-size=200`;
+
+      fetch(sectionUrl)
+        .then(res => res.json())
+        .then(packet => {
+          const response = packet.response;
+
+          if (response && response.status === 'ok') {
+            const data = response.results;
+
+            if (data && data.length) {
+              processSectionResults(data);
+              return data;
+            }
+          }
+          return false;
+        })
+        .then(data => {
+          const promises: Promise<RawTagIdData | RawSectionData>[] = [];
+
+          if (data) {
+            const sectionIds: string[] = data.map((d: RawSectionData) => {
+              return d.id;
+            });
+
+            let negativeSections = '';
+
+            sectionIds.forEach(s => {
+              const tagIdUrl = `${window.location.origin}/capi/tags?page-size=1000&section=${s}`;
+              negativeSections += `-${s},`;
+              promises.push(generateFetchPromise(tagIdUrl, promises));
+            });
+
+            // Attempt to capture tags with no set section value
+            const noSectionsUrl = `${
+              window.location.origin
+            }/capi/tags?page-size=1000&section=${negativeSections.slice(0, -1)}`;
+            promises.push(generateFetchPromise(noSectionsUrl, promises));
+          }
+
+          Promise.allSettled(promises)
+            .then(() => {
+              // Delay by a second
+              // - to capture returns from multi-paged results which take time to settle
+              setTimeout(() => {
+                processTagIdResults(rawTagIdResults);
+                console.log(`CAPI fetch completed!`);
+              }, 2000);
+            })
+            .catch(error => console.log(error));
+        })
+        .catch(error => console.log(error));
+    } else {
+      console.log('CAPI fetch either underway or completed');
+    }
+  };
+
+  callCAPI();
+
+  // To suppress opening the tagId selectors until the user has typed in 3+ characters
   const [inputValueForTagId, setInputValueForTagId] = React.useState('');
   const [openForTagId, setOpenForTagId] = React.useState(false);
-  const handleOpenForTagId = () => {
+  const handleOpenForTagId = (): void => {
     if (inputValueForTagId.length > 2) {
       setOpenForTagId(true);
     }
   };
-  const handleInputChangeForTagId = (event:any, newInputValue:string) => {
+  const handleInputChangeForTagId = (event: ChangeEvent<{}>, newInputValue: string): void => {
     setInputValueForTagId(newInputValue);
     if (newInputValue.length > 2) {
       setOpenForTagId(true);
@@ -144,12 +336,15 @@ const EpicTestTargetContentEditor: React.FC<EpicTestTargetContentEditorProps> = 
 
   const [inputValueForExcludeTagId, setInputValueForExcludeTagId] = React.useState('');
   const [openForExcludeTagId, setOpenForExcludeTagId] = React.useState(false);
-  const handleOpenForExcludeTagId = () => {
+  const handleOpenForExcludeTagId = (): void => {
     if (inputValueForExcludeTagId.length > 2) {
       setOpenForExcludeTagId(true);
     }
   };
-  const handleInputChangeForExcludeTagId = (event:any, newInputValue:string) => {
+  const handleInputChangeForExcludeTagId = (
+    event: ChangeEvent<{}>,
+    newInputValue: string,
+  ): void => {
     setInputValueForExcludeTagId(newInputValue);
     if (newInputValue.length > 2) {
       setOpenForExcludeTagId(true);
@@ -160,86 +355,72 @@ const EpicTestTargetContentEditor: React.FC<EpicTestTargetContentEditorProps> = 
 
   return (
     <div className={classes.container}>
-
       <Autocomplete
         multiple
-        options={tagIdOptions}
+        options={tagIdsOptions}
         disabled={!editMode}
-        groupBy={(opt) => opt.section}
-        getOptionLabel={(opt) => opt.display}
+        groupBy={(opt): string => opt.section}
+        getOptionLabel={(opt): string => opt.display}
         filterSelectedOptions
-        onChange={(event, value) => setTagIdsField(value)}
+        onChange={(event, value): void => setTagIdsField(value)}
         value={tagIdsField}
         onBlur={onSubmit}
         open={openForTagId}
         onOpen={handleOpenForTagId}
-        onClose={() => setOpenForTagId(false)}
+        onClose={(): void => setOpenForTagId(false)}
         inputValue={inputValueForTagId}
         onInputChange={handleInputChangeForTagId}
-        renderInput={(params) => (
-          <TextField
-            {...params}
-            name="tagIds"
-            variant="outlined"
-            label="Target tags"
-          />
+        // freeSolo
+        renderInput={(params): JSX.Element => (
+          <TextField {...params} name="tagIds" variant="outlined" label="Target tags" />
         )}
       />
 
       <Autocomplete
         multiple
-        options={sectionOptions}
+        options={sectionsOptions}
         disabled={!editMode}
-        getOptionLabel={(opt) => opt.display}
+        getOptionLabel={(opt): string => opt.display}
         filterSelectedOptions
-        onChange={(event, value) => setSectionsField(value)}
+        onChange={(event, value): void => setSectionsField(value)}
         value={sectionsField}
         onBlur={onSubmit}
-        renderInput={(params) => (
-          <TextField
-            {...params}
-            name="sections"
-            variant="outlined"
-            label="Target sections"
-          />
+        renderInput={(params): JSX.Element => (
+          <TextField {...params} name="sections" variant="outlined" label="Target sections" />
         )}
       />
 
       <Autocomplete
         multiple
-        options={tagIdOptions}
+        options={tagIdsOptions}
         disabled={!editMode}
-        groupBy={(opt) => opt.section}
-        getOptionLabel={(opt) => opt.display}
+        groupBy={(opt): string => opt.section}
+        getOptionLabel={(opt): string => opt.display}
         filterSelectedOptions
-        onChange={(event, value) => setExcludeTagIdsField(value)}
+        onChange={(event, value): void => setExcludeTagIdsField(value)}
         value={excludeTagIdsField}
         onBlur={onSubmit}
         open={openForExcludeTagId}
         onOpen={handleOpenForExcludeTagId}
-        onClose={() => setOpenForExcludeTagId(false)}
+        onClose={(): void => setOpenForExcludeTagId(false)}
         inputValue={inputValueForExcludeTagId}
         onInputChange={handleInputChangeForExcludeTagId}
-        renderInput={(params) => (
-          <TextField
-            {...params}
-            name="excludeTagIds"
-            variant="outlined"
-            label="Excluded tags"
-          />
+        // freeSolo
+        renderInput={(params): JSX.Element => (
+          <TextField {...params} name="excludeTagIds" variant="outlined" label="Excluded tags" />
         )}
       />
 
       <Autocomplete
         multiple
-        options={sectionOptions}
+        options={sectionsOptions}
         disabled={!editMode}
-        getOptionLabel={(opt) => opt.display}
+        getOptionLabel={(opt): string => opt.display}
         filterSelectedOptions
-        onChange={(event, value) => setExcludeSectionsField(value)}
+        onChange={(event, value): void => setExcludeSectionsField(value)}
         value={excludeSectionsField}
         onBlur={onSubmit}
-        renderInput={(params) => (
+        renderInput={(params): JSX.Element => (
           <TextField
             {...params}
             name="excludeSections"
